app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
redis==5.0.1